Move polygon_supernet_* modules to a separate folder
