+++ apps/indexer/lib/indexer/supervisor.ex
@@ -26,10 +26,6 @@ defmodule Indexer.Supervisor do
     PendingBlockOperationsSanitizer,
     PendingTransaction,
     PolygonSupernet,
-    PolygonSupernetDeposit,
-    PolygonSupernetDepositExecute,
-    PolygonSupernetWithdrawal,
-    PolygonSupernetWithdrawalExit,
     ReplacedTransaction,
     Token,
     TokenBalance,
@@ -128,12 +124,12 @@ defmodule Indexer.Supervisor do
          [[json_rpc_named_arguments: json_rpc_named_arguments, memory_monitor: memory_monitor]]},
         {ReplacedTransaction.Supervisor, [[memory_monitor: memory_monitor]]},
         {PolygonSupernet.Supervisor, [[memory_monitor: memory_monitor]]},
-        {PolygonSupernetDeposit.Supervisor, [[memory_monitor: memory_monitor]]},
-        {PolygonSupernetDepositExecute.Supervisor,
+        {Indexer.Fetcher.PolygonSupernet.Deposit.Supervisor, [[memory_monitor: memory_monitor]]},
+        {Indexer.Fetcher.PolygonSupernet.DepositExecute.Supervisor,
          [[memory_monitor: memory_monitor, json_rpc_named_arguments: json_rpc_named_arguments]]},
-        {PolygonSupernetWithdrawal.Supervisor,
+        {Indexer.Fetcher.PolygonSupernet.Withdrawal.Supervisor,
          [[memory_monitor: memory_monitor, json_rpc_named_arguments: json_rpc_named_arguments]]},
-        {PolygonSupernetWithdrawalExit.Supervisor, [[memory_monitor: memory_monitor]]},
+        {Indexer.Fetcher.PolygonSupernet.WithdrawalExit.Supervisor, [[memory_monitor: memory_monitor]]},
 
         # Out-of-band fetchers
         {EmptyBlocksSanitizer.Supervisor, [[json_rpc_named_arguments: json_rpc_named_arguments]]},
